scripts/requirements.txt
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
pycryptodomex==3.17
python-dotenv==1.0.0
requests==2.30.0
urllib3==2.0.2
websockets==11.0.3
youtube-dl==2021.12.17
yt-dlp==2023.3.4
